+++ src/dp_emission.c
@@ -17,23 +17,24 @@ struct dp_emission_chunk
     uint16_t* offset;
 };
 
-static inline uint16_t offset_size(uint16_t nstates) { return nstates + 1; }
-static inline uint16_t score_size(struct dp_emission const* emission, uint16_t nstates);
+static inline uint_fast16_t offset_size(uint_fast16_t nstates) { return nstates + 1; }
+static inline uint_fast16_t score_size(struct dp_emission const* emission, uint_fast16_t nstates);
 
 struct dp_emission const* dp_emission_create(struct seq_code const*      seq_code,
-                                             struct mstate const* const* mstates, uint16_t nstates)
+                                             struct mstate const* const* mstates,
+                                             uint_fast16_t               nstates)
 {
     struct dp_emission* emiss_tbl = malloc(sizeof(*emiss_tbl));
 
     emiss_tbl->offset = malloc(sizeof(*emiss_tbl->offset) * offset_size(nstates));
     emiss_tbl->offset[0] = 0;
 
-    uint16_t size = seq_code_size(seq_code, imm_state_min_seq(mstate_get_state(mstates[0])));
-    for (uint16_t i = 1; i < nstates; ++i) {
-        emiss_tbl->offset[i] = size;
+    uint_fast16_t size = seq_code_size(seq_code, imm_state_min_seq(mstate_get_state(mstates[0])));
+    for (uint_fast16_t i = 1; i < nstates; ++i) {
+        emiss_tbl->offset[i] = (uint16_t)size;
         size += seq_code_size(seq_code, imm_state_min_seq(mstate_get_state(mstates[i])));
     }
-    emiss_tbl->offset[nstates] = size;
+    emiss_tbl->offset[nstates] = (uint16_t)size;
 
     emiss_tbl->score = malloc(sizeof(*emiss_tbl->score) * score_size(emiss_tbl, nstates));
 
@@ -42,7 +43,7 @@ struct dp_emission const* dp_emission_create(struct seq_code const*      seq_cod
     uint_fast8_t          set_size = imm_abc_length(abc);
     struct imm_cartes*    cartes = imm_cartes_create(set, set_size, seq_code_max_seq(seq_code));
 
-    for (uint16_t i = 0; i < nstates; ++i) {
+    for (uint_fast16_t i = 0; i < nstates; ++i) {
 
         uint8_t min_seq = imm_state_min_seq(mstate_get_state(mstates[i]));
         for (uint8_t len = min_seq; len <= imm_state_max_seq(mstate_get_state(mstates[i])); ++len) {
@@ -52,7 +53,7 @@ struct dp_emission const* dp_emission_create(struct seq_code const*      seq_cod
             while ((item = imm_cartes_next(cartes)) != NULL) {
 
                 struct imm_seq seq = IMM_SEQ(abc, item, len);
-                uint16_t       j = seq_code_encode(seq_code, &seq);
+                uint_fast16_t  j = seq_code_encode(seq_code, &seq);
                 j -= seq_code_offset(seq_code, min_seq);
                 float score = (float)imm_state_lprob(mstate_get_state(mstates[i]), &seq);
                 emiss_tbl->score[emiss_tbl->offset[i] + j] = score;
@@ -71,11 +72,11 @@ void dp_emission_destroy(struct dp_emission const* emission)
     free_c(emission);
 }
 
-void dp_emission_offsets_dump(struct dp_emission const* emission, uint16_t nstates)
+void dp_emission_offsets_dump(struct dp_emission const* emission, uint_fast16_t nstates)
 {
     printf("state,offset\n");
-    for (uint16_t i = 0; i < nstates; ++i) {
-        printf("%" PRIu16 ",%" PRIu16 "\n", i, emission->offset[i]);
+    for (uint_fast16_t i = 0; i < nstates; ++i) {
+        printf("%" PRIu16 ",%" PRIu16 "\n", (uint16_t)i, (uint16_t)emission->offset[i]);
     }
 }
 
@@ -129,19 +130,19 @@ err:
     return NULL;
 }
 
-void dp_emission_scores_dump(struct dp_emission const* emission, uint16_t nstates)
+void dp_emission_scores_dump(struct dp_emission const* emission, uint_fast16_t nstates)
 {
     printf("state,score\n");
-    for (uint16_t i = 0; i < nstates; ++i) {
-        printf("%" PRIu16 ",%f\n", i, emission->score[i]);
+    for (uint_fast16_t i = 0; i < nstates; ++i) {
+        printf("%" PRIu16 ",%f\n", (uint16_t)i, emission->score[i]);
     }
 }
 
-int dp_emission_write(struct dp_emission const* emission, uint16_t nstates, FILE* stream)
+int dp_emission_write(struct dp_emission const* emission, uint_fast16_t nstates, FILE* stream)
 {
-    struct dp_emission_chunk chunk = {.score_size = score_size(emission, nstates),
+    struct dp_emission_chunk chunk = {.score_size = (uint16_t)score_size(emission, nstates),
                                       .score = emission->score,
-                                      .offset_size = offset_size(nstates),
+                                      .offset_size = (uint16_t)offset_size(nstates),
                                       .offset = emission->offset};
 
     if (fwrite(&chunk.score_size, sizeof(chunk.score_size), 1, stream) < 1) {
@@ -169,7 +170,7 @@ int dp_emission_write(struct dp_emission const* emission, uint16_t nstates, FILE
     return 0;
 }
 
-static inline uint16_t score_size(struct dp_emission const* emission, uint16_t nstates)
+static inline uint_fast16_t score_size(struct dp_emission const* emission, uint_fast16_t nstates)
 {
     return emission->offset[nstates];
 }
