+++ src/dp_matrix.c
@@ -9,12 +9,12 @@ struct dp_matrix* dp_matrix_create(struct dp_state_table const* states)
     matrix->states = states;
     matrix->state_col = malloc(sizeof(*matrix->state_col) * dp_state_table_nstates(states));
 
-    uint16_t next_col = 0;
-    for (uint16_t i = 0; i < dp_state_table_nstates(states); ++i) {
-        uint8_t const min = dp_state_table_min_seq(states, i);
-        uint8_t const max = dp_state_table_max_seq(states, i);
+    uint_fast16_t next_col = 0;
+    for (uint_fast16_t i = 0; i < dp_state_table_nstates(states); ++i) {
+        uint_fast8_t const min = dp_state_table_min_seq(states, i);
+        uint_fast8_t const max = dp_state_table_max_seq(states, i);
         matrix->state_col[i] = (int16_t)(next_col - min);
-        next_col += max - min + 1;
+        next_col += (uint_fast16_t)(max - min + 1);
     }
 
     matrix->score = matrixf_create(1, next_col);
